fixes main form. needs to fix styling.

+++ SeizureTracker/Context/Entities/Seizure.cs
@@ -6,7 +6,7 @@ public partial class Seizure
     public DateTime? TimeOfSeizure { get; set; }
     public string? AmPm { get; set; }
     public int? SeizureStrength { get; set; }
-    public float? KetonesLevel { get; set; }
+    public decimal? KetonesLevel { get; set; }
     public string? SeizureType { get; set; }
     public int? SleepAmount { get; set; }
     public string? Notes { get; set; }
